# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Orange Money",
        "description": "API complète pour le système de paiement Orange Money avec authentification OTP et mot de passe",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Obtenir les transactions du compte connecté",
                "description": "Récupère toutes les transactions du compte de l'utilisateur connecté avec pagination et tri par ordre décroissant",
                "operationId": "aa076a22c498e5038cfdba8e561351d6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page pour la pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de transaction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "paiement",
                                "transfert"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "reussie",
                                "echouee"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "compte_id": {
                                            "type": "integer"
                                        },
                                        "transactions": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "reference": {
                                                                "type": "string"
                                                            },
                                                            "type": {
                                                                "type": "string"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "format": "decimal"
                                                            },
                                                            "destinataire": {
                                                                "type": "string"
                                                            },
                                                            "statut": {
                                                                "type": "string"
                                                            },
                                                            "date_execution": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "filtres_appliques": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token d'accès Bearer obtenu via OTP ou authentification par mot de passe",
                "scheme": "bearer"
            }
        }
    },
...
}